Integrate frontend and backend

+++ src/main/java/com/justcountit/user/ForMeMetadata.java
@@ -1,4 +1,4 @@
 package com.justcountit.user;
 
-public record ForMeMetadata(int id ,Long from, Double price) {
+public record ForMeMetadata(Long id ,Long from, Double price) {
 }
